+++ actioncable/app/javascript/action_cable/connection.js
@@ -1,7 +1,7 @@
 import adapters from "./adapters"
 import ConnectionMonitor from "./connection_monitor"
 import INTERNAL from "./internal"
-import { log } from "./logger"
+import logger from "./logger"
 
 // Encapsulate the cable connection held by the consumer. This is an internal class not intended for direct user manipulation.
 
@@ -30,10 +30,10 @@ class Connection {
 
   open() {
     if (this.isActive()) {
-      log(`Attempted to open WebSocket, but existing socket is ${this.getState()}`)
+      logger.log(`Attempted to open WebSocket, but existing socket is ${this.getState()}`)
       return false
     } else {
-      log(`Opening WebSocket, current state is ${this.getState()}, subprotocols: ${protocols}`)
+      logger.log(`Opening WebSocket, current state is ${this.getState()}, subprotocols: ${protocols}`)
       if (this.webSocket) { this.uninstallEventHandlers() }
       this.webSocket = new adapters.WebSocket(this.consumer.url, protocols)
       this.installEventHandlers()
@@ -48,15 +48,15 @@ class Connection {
   }
 
   reopen() {
-    log(`Reopening WebSocket, current state is ${this.getState()}`)
+    logger.log(`Reopening WebSocket, current state is ${this.getState()}`)
     if (this.isActive()) {
       try {
         return this.close()
       } catch (error) {
-        log("Failed to reopen WebSocket", error)
+        logger.log("Failed to reopen WebSocket", error)
       }
       finally {
-        log(`Reopening WebSocket in ${this.constructor.reopenDelay}ms`)
+        logger.log(`Reopening WebSocket in ${this.constructor.reopenDelay}ms`)
         setTimeout(this.open, this.constructor.reopenDelay)
       }
     } else {
@@ -134,16 +134,16 @@ Connection.prototype.events = {
   },
 
   open() {
-    log(`WebSocket onopen event, using '${this.getProtocol()}' subprotocol`)
+    logger.log(`WebSocket onopen event, using '${this.getProtocol()}' subprotocol`)
     this.disconnected = false
     if (!this.isProtocolSupported()) {
-      log("Protocol is unsupported. Stopping monitor and disconnecting.")
+      logger.log("Protocol is unsupported. Stopping monitor and disconnecting.")
       return this.close({allowReconnect: false})
     }
   },
 
   close(event) {
-    log("WebSocket onclose event")
+    logger.log("WebSocket onclose event")
     if (this.disconnected) { return }
     this.disconnected = true
     this.monitor.recordDisconnect()
@@ -151,7 +151,7 @@ Connection.prototype.events = {
   },
 
   error() {
-    log("WebSocket onerror event")
+    logger.log("WebSocket onerror event")
   }
 }